move tasks module to staging app
