+++ data_api/staging/management/commands/sync_organization_data.py
@@ -2,7 +2,7 @@ import logging
 from django.core.management import BaseCommand
 from temba_client.v2 import TembaClient
 
-from data_api.api.tasks import sync_latest_data, logger as task_logger
+from data_api.staging.tasks import sync_latest_data, logger as task_logger
 from data_api.staging.models import logger as model_logger
 from data_api.staging.utils import import_org_with_client
