+++ data_api/staging/tests/test_data_import.py
@@ -10,7 +10,7 @@ from temba_client.v2 import TembaClient
 import uuid
 
 from data_api.staging.tests.test_utils import get_api_results_from_file
-from data_api.api.tasks import fetch_entity
+from data_api.staging.tasks import fetch_entity
 from data_api.staging.models import Organization, Group, SyncCheckpoint, Channel, Contact, ChannelEvent, Field, \
     Broadcast, Campaign, Flow, CampaignEvent, Runs, Label, FlowStart, Run, Boundary, Resthook, ResthookEvent, \
     ResthookSubscriber, Message
